Swap quotes
- During the first iteration, we printed the unescaped values which let to printing invalid JavaScript characters and bad things like invisible characters.
- During the second iteration, we escaped everything, which generated valid JavaScript but you lost your emojis and chinese/cyrillic characters

In this iteration, which I hope will be the last one, we maintain the string exactly as encoded and only swap quotes. The swap quotes implementation is a bit convoluted but thanks to @RReverser and @mathiasbynens on this thread https://twitter.com/Vjeux/status/822521844842401792 we've got something that appears to be working on all the test262 tests.